Add: file upload logic and refactor out the modell interface and class to a types.ts file.
