+++ Client/src/app/app.component.html
@@ -15,13 +15,11 @@
       <span>Advanced TypeScript</span>
     </mat-toolbar>
   </mat-sidenav-content>
-  
  </mat-sidenav-container>
- <atp-events-app></atp-events-app>
  <mat-list>
   <mat-list-item> <atp-events-list></atp-events-list> </mat-list-item>
  </mat-list>
-  
+<atp-navbar></atp-navbar>
