+++ Client/src/app/components/fileupload/fileupload.component.scss
@@ -1,0 +1,1 @@
+.mat-form-field { display: block; }
