+++ Client/src/app/events/events-list/events-list.component.ts
@@ -1,4 +1,4 @@
-import { coerceNumberProperty } from '@angular/cdk/coercion';
+
 import { Component, OnInit } from '@angular/core';
 
 @Component({
