+++ Client/src/app/types.ts
@@ -1,0 +1,14 @@
+export type ImageType = string | ArrayBuffer;
+
+export interface IPictureModel {
+    Image: string | null | ArrayBuffer;
+    Name: string;
+    Description: string;
+    Tags: string;
+   }
+   export class PictureModel implements IPictureModel {
+    Image!: string ;
+    Name!: string;
+    Description!: string;
+    Tags!: string;
+   }
